Added pointers, renamed pages folder to home, passed through query string to dashboard
