+++ src/components/header/header.css
@@ -42,3 +42,7 @@
     margin: 0 auto;
     margin-bottom: .2em;
 }
+
+.logout:hover {
+    cursor: pointer;
+}
