+++ src/components/repl/repl.js
@@ -144,8 +144,8 @@ class Repl extends React.Component {
               onChange={this.onChange}
               editorDidMount={this.editorDidMount}
             />
-            <button onClick={this.saveCode}>Save Code</button>
-            <button id="runCode" onClick={this.handleSubmit}>Run Code</button>
+            <button className="submit-button" onClick={this.saveCode}>Save Code</button>
+            <button className="submit-button" id="runCode" onClick={this.handleSubmit}>Run Code</button>
                     
           </form>
